requirements.txt
accelerate==1.3.0
anthropic==0.49.0
beautifulsoup4==4.13.3
dash==2.18.2
evaluate==0.4.3
fire==0.7.0
huggingface_hub==0.26.2
isal==1.6.1
langchain==0.3.23
langchain_community==0.3.21
langchain_core==0.3.51
langchain_google_community==2.0.7
langchain_openai==0.3.12
nltk==3.8.1
numpy==2.2.4
openai==1.73.0
pandas==2.2.3
plotly==6.0.0
protobuf==6.30.2
pyrootutils==1.0.4
python-dotenv==1.1.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.6.1
scipy==1.15.2
sentence_transformers==3.3.0
setuptools==58.1.0
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
urllib3==2.4.0
yfinance==0.2.55
matplotlib
seaborn
datasets==3.5.0
sentencepiece
adjustText